+++ hal/network/ncp_client/realtek/rtl_ncp_client.cpp
@@ -51,6 +51,7 @@ extern "C" {
 
 #undef IFNAMSIZ // AMBD SDK and LWIP both define this symbol...
 #include "wlan_hal.h"
+#include "bt_intf.h"
 
 extern "C" void rtw_efuse_boot_write(void);
 
@@ -309,6 +310,8 @@ int RealtekNcpClient::connect(const char* ssid, const MacAddress& bssid, WifiSec
             const NcpClientLock lock(this);
             CHECK_TRUE(connState_ == NcpConnectionState::DISCONNECTED, SYSTEM_ERROR_INVALID_STATE);
 
+            rtlk_bt_set_gnt_bt(PTA_AUTO);
+
             LOG(INFO, "Try to connect to ssid: %s", ssid);
             rtlError = wifi_connect((char*)ssid,
                                     wifiSecurityToRtlSecurity(sec),
@@ -380,6 +383,9 @@ int RealtekNcpClient::getNetworkInfo(WifiNetworkInfo* info) {
 
 int RealtekNcpClient::scan(WifiScanCallback callback, void* data) {
     const NcpClientLock lock(this);
+
+    rtlk_bt_set_gnt_bt(PTA_AUTO);
+
     CHECK_TRUE(ncpState_ == NcpState::ON, SYSTEM_ERROR_INVALID_STATE);
     struct Context {
         WifiScanCallback callback = nullptr;
@@ -447,11 +453,16 @@ int RealtekNcpClient::scan(WifiScanCallback callback, void* data) {
     for (int i = 0; i < ctx.results.size(); i++) {
         callback(ctx.results[i], data);
     }
-    if (ctx.results.size() == 0) {
+    if (rtlError && ctx.results.size() == 0) {
         // Workaround for a weird state we might enter where the wifi driver
         // is not returning any results
         rtwRadioReset();
     }
+
+    if (connectionState() == NcpConnectionState::DISCONNECTED) {
+        rtlk_bt_set_gnt_bt(PTA_BT);
+    }
+
     return rtl_error_to_system(rtlError);
 }
 
@@ -515,6 +526,9 @@ void RealtekNcpClient::connectionState(NcpConnectionState state) {
         event.type = NcpEvent::CONNECTION_STATE_CHANGED;
         event.state = connState_;
         handler(event, conf_.eventHandlerData());
+    }
+    if (state == NcpConnectionState::DISCONNECTED) {
+        rtlk_bt_set_gnt_bt(PTA_BT);
     }
 }