+++ hal/src/nRF52840/concurrent_hal.cpp
@@ -42,7 +42,7 @@
 #include "service_debug.h"
 
 // For OpenOCD FreeRTOS support
-extern const int  __attribute__((used)) uxTopUsedPriority = configMAX_PRIORITIES;
+extern const int  __attribute__((used)) uxTopUsedPriority = configMAX_PRIORITIES - 1;
 
 // use the newer name
 typedef xTaskHandle TaskHandle_t;