+++ hal/src/rtl872x/usbd_driver.h
@@ -95,7 +95,11 @@ private:
     usb_dev_t* rtlDev_ = nullptr;
     bool setupError_ = false;
 
-    const uint8_t RTL_USBD_ISR_PRIORITY = 7;
+#if MODULE_FUNCTION != MOD_FUNC_BOOTLOADER
+    const uint8_t RTL_USBD_ISR_PRIORITY = OS_THREAD_PRIORITY_NETWORK - 1;
+#else
+    const uint8_t RTL_USBD_ISR_PRIORITY = 7; // Does not matter, no RTOS
+#endif // MODULE_FUNCTION != MOD_FUNC_BOOTLOADER
     __attribute__((aligned(4))) uint8_t tempBuffer_[rtl::TEMP_BUFFER_SIZE];
 
     // TODO: Validate whether these are required to be present at all times
@@ -105,7 +109,7 @@ private:
         .rx_fifo_size = USBD_MAX_RX_FIFO_SIZE,
         .nptx_fifo_size = USBD_MAX_NPTX_FIFO_SIZE,
         .ptx_fifo_size = USBD_MAX_PTX_FIFO_SIZE,
-        .speed = USB_SPEED_FULL,
+        .speed = USB_SPEED_HIGH,
         .dma_enable = 0, // ?
         .self_powered = 1,
         .isr_priority = RTL_USBD_ISR_PRIORITY,