+++ system/inc/system_version.h
@@ -213,7 +213,8 @@ extern "C" {
 #define SYSTEM_VERSION_v550RC1        SYSTEM_VERSION_RC(5, 5, 0, 1)
 #define SYSTEM_VERSION_v550         SYSTEM_VERSION_DEFAULT(5, 5, 0)
 #define SYSTEM_VERSION_v560         SYSTEM_VERSION_DEFAULT(5, 6, 0)
-#define SYSTEM_VERSION SYSTEM_VERSION_v560
+#define SYSTEM_VERSION_v570         SYSTEM_VERSION_DEFAULT(5, 7, 0)
+#define SYSTEM_VERSION SYSTEM_VERSION_v570
 
 /**
  * Previously we would set the least significant byte to 0 for the final release, but to make
@@ -391,6 +392,7 @@ extern "C" {
 #define SYSTEM_VERSION_550RC1
 #define SYSTEM_VERSION_550
 #define SYSTEM_VERSION_560
+#define SYSTEM_VERSION_570
 
 typedef struct __attribute__((packed)) SystemVersionInfo
 {